Update bootstrap.py
- Add the ability to chose packages to install
- Make test mode actually work
- Make the printouts slightly prettier

+++ bootstrap.py
@@ -16,7 +16,9 @@ def confirm(prompt='Confirm', default=False):
         return {'y': True, 'n': False}[response]
 
 def process_package(dotfile_root, dirname, test=False):
-    print(dirname)
+    print('+--------')
+    print('| ' + dirname + (' (dryrun)' if test else ''))
+    print('+--------')
     for src in glob.iglob(os.path.join(dotfile_root, dirname, '*.symlink*')):
         filename = os.path.basename(src)
         basename, variant = filename.rsplit('.symlink', 1)
@@ -27,7 +29,7 @@ def process_package(dotfile_root, dirname, test=False):
         elif variant == 'xdg':
             xch = os.environ.get('XDG_CONFIG_HOME', None)
             if xch is None:
-                print('Skipping ' + src)
+                print('| Skipping ' + src)
                 continue
             dst = os.path.join(xch, dirname, basename)
         else:
@@ -43,29 +45,40 @@ def process_package(dotfile_root, dirname, test=False):
             if confirm(prompt):
                 try:
                     if not os.path.islink(dst) and os.path.isdir(dst):
-                        shutil.rmtree(dst)
+                        print('| rm -rf ' + dst)
+                        if not test:
+                            shutil.rmtree(dst)
                     else:
-                        os.remove(dst)
+                        print('| Would rm ' + dst)
+                        if not test:
+                            os.remove(dst)
                 except OSError:
                     pass
             else:
-                print('Skipping ' + dst)
+                print('| Skipping ' + dst)
                 continue
 
         parent = os.path.dirname(dst)
         if not os.path.exists(parent):
-            os.makedirs(parent)
+            print('| mkdir ' + parent)
+            if not test:
+                os.makedirs(parent)
 
-        os.symlink(src, dst)
+        print('| Symlink ' + dst + ' -> ' + src)
+        if not test:
+            os.symlink(src, dst)
+    print('+--------')
 
-def main(test=False):
+def main(packages, test=False):
+    packages = set(packages)
     dotfile_root = os.path.abspath(os.path.dirname(__file__))
     for dirname in os.listdir(dotfile_root):
         if not os.path.isdir(os.path.join(dotfile_root, dirname)):
             continue
         if dirname in ('.git',):
             continue
-        process_package(dotfile_root, dirname, test=test)
+        if not packages or dirname in packages:
+            process_package(dotfile_root, dirname, test=test)
 
 if __name__ == '__main__':
     # Use optparse to support old vanilla pythons
@@ -73,6 +86,4 @@ if __name__ == '__main__':
     parser = optparse.OptionParser()
     parser.add_option('-n', '--test', action='store_true')      
     opts, args = parser.parse_args()
-    if len(args) != 0:
-        parser.error('Incorrect number of arguments! (expect 0, got {0})'.format(len(args)))
-    main(**vars(opts))
+    main(args, **vars(opts))
